Project init

+++ statistic-service/src/main/java/ru/practicum/statservice/server/StatCountHits.java
@@ -13,6 +13,5 @@ public class StatCountHits {
 
     private String uri;
 
-    private Long hits;
-
+    private Integer hits;
 }
